Rewrite using SwiftUI (fixes #1)
The diff is larger than it may need to be, but I scrapped the whole
project and re-created it in SwiftUI. Localizations are missing and I
noticed some other differences, such as default menu items being
present.

The benefit of scrapping the whole project is that you get more modern
project configuration and the new localization formats.