Remove some superfluous menu entries (fixes #5)
Note that only a few selected menu entries can be removed in SwiftUI by
default, or all of them (including "Quit"). To avoid hacking the menu
through `NSApplication`, let's keep it at that and see if more options
are present in the future. There are multiple questions on StackOverflow
and in Apple's developer forums and people already reported this as
bugs.