Add command-line interface (fixes #4)
The change does not add a separate command-line utility, but just a
command-line interface to the default executable. There are three main
reasons for this:

1. Only the executable of the containing application has permissions to
   interact with the extension. I assume it is a sandboxing rule, but I
   have not confirmed this explicitly. If you try to interact with the
   extension, it will not be found. Renaming the executable and
   replacing the containing app with that executable however works.
2. The application groups entitlement needs to be added explicitly and
   the sandbox logs that this will be ignored. However, I have not
   confired whether this is an actual problem.
3. The bundle identifier of the containing application is a prefix of
   the extension. It is unclear how that effects command-line utilities
   and whether it can be worked around by having the containing
   application's bundle identifier be a prefix of the command-line
   utility, which then in turn could be a prefix of the extension.

All of these issues are avoided by extending the main executable of the
application with additional command-line options. The UI will be
launched by default, to not conflict with the default behaviour when
staring the application.

Note that applications accept some command-line arguments by default,
such as `-ApplePersistenceIgnoreState` or `-NSDocumentRevisionsDebugMode`.
I still need to figure out how to pass through unprocessed command-line
arguments.