Pass-through system command-line arguments (see #4)
UI applications support a range of standard arguments provided by the
system, such as `-NSShowNonLocalizedStrings` to reveal localization
issues, `-ApplePersistenceIgnoreState`, to avoid persisting the
application's state, and more.

The arguments now are not treated as an error. There is no need to
distinguish them any further, since the UI application is not run
through the command-line parser itself, but simply the `main()` function
of the UI application is invoked. It seems to ignore the `ui`
sub-command, which is even better.

Note that if using the debugging options in Xcode, which would set some
of those system-provieded arguments, the other sub-commands will fail
due to unknown arguments.