Suggest using shell aliases instead of symbolic links (#10)
When launching the executable through a symbolic link,
the application is started differently, i.e., outside
of the application bundle. This disconnects the
command-line tool from the information property list
(`Info.plist`). When trying to access the main bundle
of the application the process fails to detect the
bundle identifier, which is usually obtained from the
default information property list. This causes a
forced unwrap of `nil` and crashes the process.

This does not happen if a shell alias is used instead.