Hard-code command name (#6)
Since launching through a symbolic link is not working (#10), one has to
call the exectuable directly or create an alias. All of these options
basically result in the full path to the executable being in the first
command-line argument, which is a bit verbose. Let's hard-code it after
all, to avoid bloated help output.